Minimal fixes to get Masto Feeder working with Sky Bridge
https://skybridge.fly.dev/ is an adapter that allows exports a Mastodon
API endpoint for Bluesky. Masto Feeder mostly works as is, but needed
a few tweaks:
- The "state" param in the OAuth flow was not being passed back, so
  switch to using a cookie to store the pending session state (which
  is more correct anyway, "state" is meant to be used for XSRF
  protection)
- Sky Bridge returns Mastodon API version 1, which was causing it to
  get rejected by some versioned endpoints (like list fetching). Stop
  doing version number checks for now.
- Sky Bridge did not have quite the same max_id status fetchign behavior,
  which was resulting in infinite loops. Add a guard to make sure we stop
  if we don't get back any new statuses.

+++ app/mastofeeder/handlers/session.py
@@ -10,6 +10,8 @@ import mastofeeder.data as data
 BASE_PATH = '/masto-feeder'
 SCOPES = ['read:accounts', 'read:follows', 'read:lists', 'read:statuses']
 
+AUTH_REQUEST_COOKIE_NAME = 'mastofeeder-auth-request'
+
 class BaseHandler(base.handlers.BaseHandler):
     def _get_path(self, path=''):
         return '%s/%s' % (BASE_PATH, path)
@@ -19,35 +21,47 @@ class BaseHandler(base.handlers.BaseHandler):
         return '%s://%s%s' % (
             request_url.scheme, request_url.netloc, self._get_path(path))
 
-class SessionHandler(BaseHandler):
-    SESSION_COOKIE_NAME = 'mastofeeder-sid'
+    def _has_cookie(self, name):
+        return name in self.request.cookies
 
-    def _has_request_session(self):
-        return self.SESSION_COOKIE_NAME in self.request.cookies
+    def _get_cookie(self, name):
+        return self.request.cookies[name]
 
-    def _get_session_from_request(self):
-        session_id = self.request.cookies[self.SESSION_COOKIE_NAME]
-        return data.Session.get_by_session_id(session_id)
-
-    def _set_request_session(self, session):
+    def _set_cookie(self, name, value):
         cookie = Cookie.SimpleCookie()
-        cookie[self.SESSION_COOKIE_NAME] = session.session_id
-        morsel = cookie[self.SESSION_COOKIE_NAME]
+        cookie[name] = value
+        morsel = cookie[name]
         morsel['path'] = BASE_PATH
         # TODO(mihaip): expiration?
 
         self.response.headers.add_header(
             'Set-Cookie', morsel.output(header='').lstrip())
 
-    def _remove_request_session(self):
+    def _remove_cookie(self, name):
         cookie = Cookie.SimpleCookie()
-        cookie[self.SESSION_COOKIE_NAME] = 'expired'
-        morsel = cookie[self.SESSION_COOKIE_NAME]
+        cookie[name] = 'expired'
+        morsel = cookie[name]
         morsel['path'] = BASE_PATH
         morsel['expires'] = 'Sat, 1-Jan-2000 00:00:00'
 
         self.response.headers.add_header(
             'Set-Cookie', morsel.output(header='').lstrip())
+
+class SessionHandler(BaseHandler):
+    SESSION_COOKIE_NAME = 'mastofeeder-sid'
+
+    def _has_request_session(self):
+        return self._has_cookie(self.SESSION_COOKIE_NAME)
+
+    def _get_session_from_request(self):
+        session_id = self._get_cookie(self.SESSION_COOKIE_NAME)
+        return data.Session.get_by_session_id(session_id)
+
+    def _set_request_session(self, session):
+        self._set_cookie(self.SESSION_COOKIE_NAME, session.session_id)
+
+    def _remove_request_session(self):
+        self._remove_cookie(self.SESSION_COOKIE_NAME)
 
     def _get_feed_path(self, *args):
         return self._get_path('feed/%s/%s' % (
@@ -96,6 +110,7 @@ class SessionApiHandler(SessionHandler):
         raise NotImplementedError()
 
 class SignInHandler(BaseHandler):
+
     def post(self):
         instance_url = self.request.get('instance_url')
 
@@ -124,6 +139,7 @@ class SignInHandler(BaseHandler):
         auth_request_data = data.MastodonAuthRequest.create(
             instance_url=instance_url)
         auth_request_data.put()
+        self._set_cookie(AUTH_REQUEST_COOKIE_NAME, auth_request_data.id)
 
         auth_request_url = api.auth_request_url(
             scopes=SCOPES,
@@ -137,11 +153,20 @@ class SignInHandler(BaseHandler):
 class SignInCallbackHandler(SessionHandler):
     def get(self):
         code = self.request.get('code')
+        auth_request_id = self._get_cookie(AUTH_REQUEST_COOKIE_NAME)
+        if not auth_request_id:
+            self._write_input_error('No auth request cookie found')
+            return
+        self._remove_cookie(AUTH_REQUEST_COOKIE_NAME)
         state = self.request.get('state')
+        # Sky Bridge does not send back the state parameter.
+        if state and state != auth_request_id:
+            self._write_input_error('Mismatched auth request cookie (%s) and state (%s)' % (auth_request_id, code))
+            return
 
-        auth_request_data = data.MastodonAuthRequest.get_by_id(state)
+        auth_request_data = data.MastodonAuthRequest.get_by_id(auth_request_id)
         if not auth_request_data:
-            self._write_input_error('Unknown auth request %s' % code)
+            self._write_input_error('Unknown auth request %s' % auth_request_id)
             return
         instance_url = auth_request_data.instance_url
         auth_request_data.delete()
